Remove usage of offset and rely only id condition instead.
To keep encryptable batches self-sufficient,
we calculate the next id by fetching the next ids in the minimal amount
we need to do it. The resulting query will be as big as number of tracks *
size per batch, we we consider an acceptable tradeoff.

Alternatively, we would have to introduce some kind of external knowledge about
how to fetch the next page with the corresponding offset. A given batch
doesn't have a direct way of fetching the last id in the preceding batch inside
a track group.

offset is very slow in MySQL for large values. See conversation with @lewispb in
https://github.com/basecamp/mass_encryption/pull/3#issuecomment-936301911.